requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2

sqlalchemy==2.0.32
alembic==1.13.2
pydantic==1.10.13
python-dotenv==1.0.1

python-telegram-bot==21.4
requests==2.32.3
reportlab==4.2.2
redis==5.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
pyotp==2.9.0
boto3==1.37.0
cryptography==44.0.2
psycopg2-binary==2.9.10

pytest==8.3.4
httpx==0.27.2

# --- Senior IT: Distributed Tracing & Quality ---
structlog==25.1.0
ruff==0.9.7
mypy==1.15.0
holidays==0.67
slowapi==0.1.9
bandit==1.8.3
diskcache==5.6.3
psutil==5.9.8
vulture==2.11
pydeps==1.12.19
sentry-sdk[fastapi]==2.8.0
prometheus_client==0.20.0
